+++ src/core/webview/ClineProvider.ts
@@ -680,7 +680,7 @@ export class ClineProvider extends EventEmitter<ClineProviderEvents> implements 
 						window.AUDIO_BASE_URI = "${audioUri}"
 						window.MATERIAL_ICONS_BASE_URI = "${materialIconsUri}"
 					</script>
-					<title>Shenma</title>
+					<title>Costrict</title>
 				</head>
 				<body>
 					<div id="root"></div>
@@ -776,7 +776,7 @@ export class ClineProvider extends EventEmitter<ClineProviderEvents> implements 
 				window.AUDIO_BASE_URI = "${audioUri}"
 				window.MATERIAL_ICONS_BASE_URI = "${materialIconsUri}"
 			</script>
-            <title>Shenma</title>
+            <title>Costrict</title>
           </head>
           <body>
             <noscript>You need to enable JavaScript to run this app.</noscript>