+++ src/integrations/terminal/__tests__/TerminalRegistry.test.ts
@@ -34,7 +34,7 @@ describe("TerminalRegistry", () => {
 
 			expect(mockCreateTerminal).toHaveBeenCalledWith({
 				cwd: "/test/path",
-				name: "Shenma",
+				name: "Costrict",
 				iconPath: expect.any(Object),
 				env: {
 					PAGER: "cat",
@@ -54,7 +54,7 @@ describe("TerminalRegistry", () => {
 
 				expect(mockCreateTerminal).toHaveBeenCalledWith({
 					cwd: "/test/path",
-					name: "Shenma",
+					name: "Costrict",
 					iconPath: expect.any(Object),
 					env: {
 						PAGER: "cat",
@@ -76,7 +76,7 @@ describe("TerminalRegistry", () => {
 
 				expect(mockCreateTerminal).toHaveBeenCalledWith({
 					cwd: "/test/path",
-					name: "Shenma",
+					name: "Costrict",
 					iconPath: expect.any(Object),
 					env: {
 						PAGER: "cat",