+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Shenma Change Log
+# Costrict Change Log
 
 ## [1.5.6]
 - Update Error Code